feat(ask-core): F0 slice 2 — extract the streaming protocol; split client/Node entries (RFC-0007)
Move lib/ask/protocol.ts (wire events + generative-UI tool defs, corpus-agnostic +
client-safe) into @agentskit/ask-core's default entry. Embedder moves to a separate
@agentskit/ask-core/embed entry so the browser widget importing the protocol never
pulls @huggingface/transformers. docs-next keeps thin shims at the old paths.
